src/compression/state_delta.py
# state_delta.py

"""
Module: StateDelta
Purpose:
    The `StateDelta` module is designed to optimize state storage in blockchain systems by 
    focusing on storing only the changes (deltas) that occur as a result of transactions, 
    rather than the full state at every block. This dramatically reduces the storage overhead, 
    especially in systems with frequent transactions, by only maintaining the updates.

    In blockchain systems, every transaction changes the state of the system (e.g., account balances, 
    smart contract states). Storing the entire state after every block leads to inefficiency. By storing 
    only the deltas, we ensure that the blockchain node uses less disk space, while still being able to 
    reconstruct the full state when needed.

    Key Features:
    - Efficient storage of state changes by tracking only the deltas.
    - Real-time updates to the current state based on applied transactions.

    Security Considerations:
    The StateDelta module does not perform cryptographic operations, but in a secure blockchain system, 
    the transactions themselves must be verified and signed to ensure authenticity before they are applied 
    to the state. Additionally, delta application should be deterministic and idempotent to avoid state corruption.
"""
class StateDelta:
    """
    The `StateDelta` class manages the storage and application of deltas, which represent 
    changes in the blockchain state caused by transactions. Instead of storing the full state 
    after every transaction, it stores only the differences (deltas), making state management 
    more efficient in terms of storage space.

    Scalability Consideration:
    By only storing changes (deltas), this design optimizes memory usage, allowing blockchain 
    systems to scale without bloating the state database unnecessarily. This is particularly 
    effective in systems with high transaction throughput.
    """

    def __init__(self):
        """
        Initializes the `StateDelta` instance with an empty `current_state` dictionary.
        This dictionary will hold the cumulative state of the blockchain by tracking 
        account balances or other relevant state variables as transactions are applied.
        
        Design Consideration:
        The `current_state` is represented as a dictionary where keys are identifiers (e.g., account names) 
        and values represent the current balance or state value. This structure allows for efficient lookups 
        and updates.
        """
        self.current_state = {}  # The dictionary that holds the current state of all accounts.

    def apply_transaction(self, transaction: dict) -> dict:
        """
        Applies a transaction to the current state and generates the delta (the difference in the state).

        :param transaction: A dictionary representing a blockchain transaction.
                            Each transaction must include an 'account' field (identifier) and an 'amount' field 
                            that modifies the account balance.
        
        :return: The delta (difference) applied to the state as a dictionary.
                 The returned delta contains the account and its updated balance after applying the transaction.
        
        Process:
        - Retrieve the current balance (or default to zero if the account is new).
        - Update the balance by adding the transaction amount.
        - Store the change in the `current_state` dictionary and return the delta.
        
        Security Note:
        This function assumes that the transaction has already been validated by other mechanisms in the 
        blockchain system (e.g., signature verification). It should not be called with unverified data to 
        avoid state corruption.
        """
        delta = {}  # A dictionary that will store the delta change in the state for this transaction.
        
        # Retrieve the account and its current balance (or assume zero if the account doesn't exist).
        account = transaction['account']
        
        # Calculate the new balance by adding the transaction's amount to the current state.
        delta[account] = self.current_state.get(account, 0) + transaction['amount']
        
        # Update the current state with the new balance.
        self.current_state.update(delta)
        
        # Return the delta that was applied for record-keeping or further processing.
        return delta

    def get_current_state(self) -> dict:
        """
        Returns the full current state of the blockchain, represented as a dictionary.
        
        :return: The entire current state as a dictionary, where each key is an account identifier 
                 and the value is the respective balance.
        
        Use Case:
        This method is useful when the full state of the system is required, such as during validation, 
        reporting, or auditing.
        
        Performance Consideration:
        Although the state is updated incrementally using deltas, accessing the full current state 
        may involve a large data set in long-running blockchain systems. In such cases, additional 
        indexing or state compression mechanisms may be considered for future optimizations.
        """
        return self.current_state  # Return the entire current state as a dictionary.
"""
Security and Architectural Considerations:

1. **Data Integrity**:
   The `StateDelta` class ensures that only validated transactions are applied to the state. 
   However, it is crucial that transactions are authenticated and verified (e.g., with cryptographic 
   signatures) before they are passed to this module. This ensures the state remains consistent and secure.

2. **Scalability**:
   By focusing on delta-based storage, this module minimizes the memory and disk space required 
   to store state data in large blockchain systems. This approach reduces the overall data footprint 
   and enhances performance when dealing with frequent state changes.

3. **Consistency**:
   The `apply_transaction` method is designed to be idempotent, meaning that applying the same 
   transaction multiple times should result in the same state. This property ensures consistency 
   even in cases of network interruptions or retries in distributed systems.
"""
